+++ apps/api/test/seeders/user.seeder.ts
@@ -19,7 +19,8 @@ export class UserSeeder {
     lastIp = faker.internet.ip(),
     lastUserAgent = faker.internet.userAgent(),
     lastFingerprint = faker.word.noun(),
-    createdAt = faker.date.recent()
+    createdAt = faker.date.recent(),
+    trial = false
   }: Partial<UserOutput> = {}): UserOutput {
     return {
       id,
@@ -37,7 +38,8 @@ export class UserSeeder {
       lastIp,
       lastUserAgent,
       lastFingerprint,
-      createdAt
+      createdAt,
+      trial
     };
   }
 }