+++ apps/log-collector/src/providers/process-env.provider.ts
@@ -1,6 +1,0 @@
-import process from "node:process";
-import { container } from "tsyringe";
-
-export const PROCESS_ENV = "PROCESS_ENV";
-
-container.register(PROCESS_ENV, { useValue: process.env });