+++ apps/log-collector/CHANGELOG.md
@@ -1,5 +1,36 @@
 
 
+## [1.7.0](https://github.com/akash-network/console/compare/log-collector/v1.6.0...log-collector/v1.7.0) (2025-08-26)
+
+
+### Features
+
+* adds api background-jobs server setup ([#1833](https://github.com/akash-network/console/issues/1833)) ([d3e6214](https://github.com/akash-network/console/commit/d3e6214800722fafd872a876ddaff0591a6e6dd8))
+
+## [1.6.0](https://github.com/akash-network/console/compare/log-collector/v1.5.0...log-collector/v1.6.0) (2025-08-25)
+
+
+### Features
+
+* make log outputs configurable and add local k8s dev setup ([#1832](https://github.com/akash-network/console/issues/1832)) ([754a3cb](https://github.com/akash-network/console/commit/754a3cb4ece83604d6391436b48f6931f9ecfb1f))
+
+
+### Bug Fixes
+
+* **deployment:** handles invalid manifest errors on POST /v1/leases ([f5da5c4](https://github.com/akash-network/console/commit/f5da5c4b02ef3e2977a8f5855eb5a8b81ac8281b)), closes [#1835](https://github.com/akash-network/console/issues/1835)
+
+## [1.5.0](https://github.com/akash-network/console/compare/log-collector/v1.4.0...log-collector/v1.5.0) (2025-08-21)
+
+
+### Features
+
+* **log-collector:** implements fs log streaming and delivery with fluent-bit ([c05fe61](https://github.com/akash-network/console/commit/c05fe61bfbe9d218f9c88c1d0e1b3c74ec4a5d64))
+
+
+### Bug Fixes
+
+* refactor http-sdk services to accept httpClient ([#1829](https://github.com/akash-network/console/issues/1829)) ([abcb7dc](https://github.com/akash-network/console/commit/abcb7dc9eaeca626e6ba69edb561ff0172cf6c1a))
+
 ## [1.4.0](https://github.com/akash-network/console/compare/log-collector/v1.3.1...log-collector/v1.4.0) (2025-08-14)