+++ apps/provider-console/CHANGELOG.md
@@ -1,5 +1,19 @@
 
 
+## [1.22.0](https://github.com/akash-network/console/compare/provider-console/v1.21.1...provider-console/v1.22.0) (2025-08-26)
+
+
+### Features
+
+* adds api background-jobs server setup ([#1833](https://github.com/akash-network/console/issues/1833)) ([d3e6214](https://github.com/akash-network/console/commit/d3e6214800722fafd872a876ddaff0591a6e6dd8))
+
+## [1.21.1](https://github.com/akash-network/console/compare/provider-console/v1.21.0...provider-console/v1.21.1) (2025-08-25)
+
+
+### Bug Fixes
+
+* **deployment:** handles invalid manifest errors on POST /v1/leases ([f5da5c4](https://github.com/akash-network/console/commit/f5da5c4b02ef3e2977a8f5855eb5a8b81ac8281b)), closes [#1835](https://github.com/akash-network/console/issues/1835)
+
 ## [1.21.0](https://github.com/akash-network/console/compare/provider-console/v1.20.0...provider-console/v1.21.0) (2025-08-18)